Renamed CCR with _rules #712
