+++ castervoice/rules/ccr/bash_rules/bash.py
