+++ castervoice/rules/ccr/recording_rules/alias/simple_alias.py
@@ -1,5 +1,5 @@
 from castervoice.lib import settings
-from castervoice.rules.ccr.recording.alias.base_alias import BaseAliasRule
+from castervoice.rules.ccr.recording_rules.alias.base_alias import BaseAliasRule
 from castervoice.lib.ctrl.mgr.rule_details import RuleDetails
