+++ castervoice/rules/ccr/vhdl_rules/vhdl2.py
@@ -7,7 +7,7 @@ try:  # Try  first loading  from caster user directory
     from vhdl_support import for_generate_string, if_generate_string, process_string, \
         case_string, component_declaration_string, component_string, architecture_string, entity_string
 except ImportError:
-    from castervoice.rules.ccr.vhdl.vhdl_support import for_generate_string, if_generate_string, process_string, \
+    from castervoice.rules.ccr.vhdl_rules.vhdl_support import for_generate_string, if_generate_string, process_string, \
         case_string, component_declaration_string, component_string, architecture_string, entity_string
 
 from castervoice.lib.ctrl.mgr.rule_details import RuleDetails
