+++ tests/rules/ccr/test_rule_modules.py
@@ -3,42 +3,42 @@ from tests.test_util.modules_testing import ModulesTestCase
 
 class CCRModulesTestCase(ModulesTestCase):
     def _rule_modules(self):
-        from castervoice.rules.ccr.bash import bash
+        from castervoice.rules.ccr.bash_rules import bash
         from castervoice.rules.core.text_manipulation_rules import text_manipulation
         from castervoice.rules.core.numbers_rules import numeric
         from castervoice.rules.core.navigation_rules import nav
         from castervoice.rules.core.navigation_rules import nav2
         from castervoice.rules.core.alphabet_rules import alphabet
         from castervoice.rules.core.punctuation_rules import punctuation
-        from castervoice.rules.ccr.cpp import cpp
-        from castervoice.rules.ccr.csharp import csharp
-        from castervoice.rules.ccr.dart import dart
-        from castervoice.rules.ccr.go import go
-        from castervoice.rules.ccr.haxe import haxe
-        from castervoice.rules.ccr.html import html_rule
-        from castervoice.rules.ccr.java import java
-        from castervoice.rules.ccr.java import java2
-        from castervoice.rules.ccr.javascript import javascript
-        from castervoice.rules.ccr.latex import latex
-        from castervoice.rules.ccr.markdown import markdown
-        from castervoice.rules.ccr.matlab import matlab
-        from castervoice.rules.ccr.matlab import matlab2
-        from castervoice.rules.ccr.prolog import prolog
-        from castervoice.rules.ccr.prolog import prolog2
-        from castervoice.rules.ccr.python import python2
-        from castervoice.rules.ccr.python import python
-        from castervoice.rules.ccr.r import r
-        from castervoice.rules.ccr.recording import bringme
-        from castervoice.rules.ccr.recording import history
-        from castervoice.rules.ccr.recording import again
-        from castervoice.rules.ccr.recording.alias import simple_alias
-        from castervoice.rules.ccr.recording.alias import chain_alias
-        from castervoice.rules.ccr.rust import rust
-        from castervoice.rules.ccr.rust import rust2
-        from castervoice.rules.ccr.sql import sql
-        from castervoice.rules.ccr.vhdl import vhdl2
-        from castervoice.rules.ccr.vhdl import vhdl
-        from castervoice.rules.ccr.voice_dev_commands import voice_dev_commands
+        from castervoice.rules.ccr.cpp_rules import cpp
+        from castervoice.rules.ccr.csharp_rules import csharp
+        from castervoice.rules.ccr.dart_rules import dart
+        from castervoice.rules.ccr.go_rules import go
+        from castervoice.rules.ccr.haxe_rules import haxe
+        from castervoice.rules.ccr.html_rules import html_rule
+        from castervoice.rules.ccr.java_rules import java
+        from castervoice.rules.ccr.java_rules import java2
+        from castervoice.rules.ccr.javascript_rules import javascript
+        from castervoice.rules.ccr.latex_rules import latex
+        from castervoice.rules.ccr.markdown_rules import markdown
+        from castervoice.rules.ccr.matlab_rules import matlab
+        from castervoice.rules.ccr.matlab_rules import matlab2
+        from castervoice.rules.ccr.prolog_rules import prolog
+        from castervoice.rules.ccr.prolog_rules import prolog2
+        from castervoice.rules.ccr.python_rules import python2
+        from castervoice.rules.ccr.python_rules import python
+        from castervoice.rules.ccr.r_rules import r
+        from castervoice.rules.ccr.recording_rules import bringme
+        from castervoice.rules.ccr.recording_rules import history
+        from castervoice.rules.ccr.recording_rules import again
+        from castervoice.rules.ccr.recording_rules.alias import simple_alias
+        from castervoice.rules.ccr.recording_rules.alias import chain_alias
+        from castervoice.rules.ccr.rust_rules import rust
+        from castervoice.rules.ccr.rust_rules import rust2
+        from castervoice.rules.ccr.sql_rules import sql
+        from castervoice.rules.ccr.vhdl_rules import vhdl2
+        from castervoice.rules.ccr.vhdl_rules import vhdl
+        from castervoice.rules.ccr.voice_dev_commands_rules import voice_dev_commands
         return [bash, alphabet, nav, nav2, numeric, punctuation,
                 text_manipulation, cpp, csharp, dart, go, haxe,
                 html_rule, java, java2, javascript, latex,
